added update quantity to product

+++ src/main/java/com/dreambeans/coffee/controller/ProductsController.java
@@ -2,6 +2,9 @@ package com.dreambeans.coffee.controller;
 
 import com.dreambeans.coffee.models.*;
 import com.dreambeans.coffee.service.*;
+
+import jakarta.servlet.http.HttpServletRequest;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
@@ -11,7 +14,7 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/products")
-public class ProductsController {
+public class ProductsController extends CookieController {
 
     private ProductService productService;
     private MenuService menuService;
@@ -21,16 +24,27 @@ public class ProductsController {
     }
 
     @GetMapping("/allproducts/{menuId}")
-    public Iterable<Product> getAllProducts(@PathVariable Long menuId) {
+    public Iterable<Product> getAllProducts(@PathVariable Long menuId, HttpServletRequest request) throws Exception {
+        if (getCookieValue(request) == null) {
+            throw new Exception("not logged in");
+        }
         return productService.listProductsByMenuId(menuId);
     }
-     @GetMapping("/allproducts")
-    public Iterable<Product> getAllProducts() {
+
+    @GetMapping("/allproducts")
+    public Iterable<Product> getAllProducts(HttpServletRequest request) throws Exception {
+        if (getCookieValue(request) == null) {
+            throw new Exception("not logged in");
+        }
         return productService.listAllProducts();
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ProductDto> getProductById(@PathVariable Long id) {
+    public ResponseEntity<ProductDto> getProductById(@PathVariable Long id, HttpServletRequest request)
+            throws Exception {
+        if (getCookieValue(request) == null) {
+            throw new Exception("not logged in");
+        }
         ProductDto productDto = null;
         Optional<Product> productOp = productService.findProductsById(id);
         if (productOp.isPresent()) {
@@ -45,7 +59,10 @@ public class ProductsController {
     }
 
     @GetMapping("/delete/{id}")
-    public ResponseEntity<?> deleteProductById(@PathVariable Long id) {
+    public ResponseEntity<?> deleteProductById(@PathVariable Long id, HttpServletRequest request) throws Exception {
+        if (getCookieValue(request) == null) {
+            throw new Exception("not logged in");
+        }
         productService.deleteProductById(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -59,7 +76,11 @@ public class ProductsController {
      * @return
      */
     @PostMapping("/save")
-    public ResponseEntity<Product> createProduct(@RequestBody ProductDto productDto) {
+    public ResponseEntity<Product> createProduct(@RequestBody ProductDto productDto, HttpServletRequest request)
+            throws Exception {
+        if (getCookieValue(request) == null) {
+            throw new Exception("not logged in");
+        }
         Menu menu = menuService.findMenuById(productDto.getMenuId()).get();
         Product product = new Product(productDto.getPrice(), productDto.getDescription(), productDto.getName(),
                 productDto.getImage(), productDto.getQuantity(), menu);
@@ -67,37 +88,27 @@ public class ProductsController {
         Product savedProduct = productService.saveProduct(product);
         return ResponseEntity.status(HttpStatus.CREATED).body(savedProduct);
     }
+/**
+ * update the quantity for the product 
+ * @param id this is the id for the product
+ * @param quantity this is the quantity 
+ * @return returns the new quantity
 
+ */
+    @GetMapping("/updateQuantity/{id}/{quantity}")
+    public ResponseEntity<Integer> updateQuantity(@PathVariable("id") int id, @PathVariable("quantity") int quantity,
+            HttpServletRequest request) throws Exception {
+        if (getCookieValue(request) == null) {
+            throw new Exception("not logged in");
+        }
+        Optional<Product> productOP = productService.findProductsById(null);
+        if (productOP.isPresent()) {
+            Product product = productOP.get();
+            product.setQuantity(quantity);
+            product = productService.saveProduct(product);
+            return ResponseEntity.status(HttpStatus.OK).body(product.getQuantity());
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
 
-
-    // @GetMapping("/product")
-    // public Product getOneProduct() {
-    // Product latte = new Product(3.99, "latte", "hot coffee",
-    // "https://www.allrecipes.com/thmb/Dq9kocJWJNmhfPotGiZT6Tl_r1w=/750x0/filters:no_upscale([…]):format(webp)/9428203-9d140a4ed1424824a7ddd358e6161473.jpg");
-    // return latte;
-    // }
-
-    // @PutMapping("/{id}")
-    // public ResponseEntity<Product> updateProduct(@PathVariable Long id,
-    // @RequestBody Product updatedProduct) {
-    // Optional<Product> existingProduct = productService.findById(id);
-    // if(existingProduct.isPresent()) {
-    // Product product = existingProduct.get();
-    // product.setName(updatedProduct.getName());
-    // product.setDescription(updatedProduct.getDescription());
-    // product.setPrice(updatedProduct.getPrice());
-    // Product savedProduct = productService.save(product);
-    // return ResponseEntity.ok(savedProduct);
-    // }else {
-    // return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-    // }
 }
-
-// @DeleteMapping("/{id}")
-// public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
-// Optional<Product> existingProduct = Repository.findById(id);
-// if (existingProduct.isPresent()) {
-// Repository.deleteById(id);
-// return ResponseEntity.noContent().build();
-// }
-// }
